Added context for is top banner open to store. Margins are correct for top banner being open with cart component.
